# COLLEGE/iiitv/SEM_3/Web_Services/social_media/backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: NamoGram RESTful Web Services API
  version: 1.0.0
  description: |
    Comprehensive RESTful API for NamoGram (Instagram-like social media platform with modern blue aesthetic).
    Fulfills all Module 2 REST requirements:
    - Resource-shaped URLs under /api/v1
    - Correct HTTP methods (GET, POST, PUT, PATCH, DELETE)
    - 8+ distinct status codes (200, 201, 204, 304, 400, 401, 403, 404, 409, 422, 429, 503)
    - One consistent RFC error body
    - Idempotency-Key support on POST
    - ETag and 304 Not Modified
    - Deliberate Cache-Control headers
    - Ajv JSON Schema validation
    - Content negotiation (application/json and application/xml)
    - Cursor pagination with next links
    - Sliding window rate limiting with headers
servers:
  - url: http://localhost:5000
    description: Local Development Server
paths:
...
  /api/v1/posts:
    get:
      summary: Retrieve posts with cursor pagination and cache control
      operationId: getPosts
      tags: [Posts]
      parameters:
        - name: limit
          in: query
          description: Max number of items to return (1-50)
          schema:
            type: integer
            default: 5
        - name: cursor
          in: query
          description: Post ID cursor for keyset pagination
          schema:
            type: string
        - name: tag
          in: query
          description: Filter posts by tag name
          schema:
            type: string
        - name: Accept
          in: header
          description: Content negotiation header (application/json or application/xml)
          schema:
            type: string
            default: application/json
      responses:
        '200':
          description: Paginated posts collection
          headers:
            ETag:
              schema: { type: string }
              description: Entity tag for conditional requests
            Cache-Control:
              schema: { type: string }
              description: Deliberate cache header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostsFeedResponse'
            application/xml:
              schema:
                type: string
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '429':
          $ref: '#/components/responses/RateLimitError'

    post:
      summary: Create a new post with schema validation and idempotency
      operationId: createPost
      tags: [Posts]
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Unique client key to guarantee idempotent creation
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreateInput'
      responses:
        '201':
          description: Post created successfully
          headers:
            Location:
              schema: { type: string }
              description: URI of newly created resource
            Idempotent-Replay:
              schema: { type: string }
              description: Returned if response is replayed from cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/posts/{id}:
    get:
      summary: Retrieve single post by ID with ETag support
      operationId: getPostById
      tags: [Posts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: If-None-Match
          in: header
          description: ETag from previous fetch for 304 conditional check
          schema: { type: string }
      responses:
        '200':
          description: Post detail
          headers:
            ETag:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      summary: Fully replace post
      operationId: replacePost
      tags: [Posts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreateInput'
      responses:
        '200':
          description: Post updated
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

    patch:
      summary: Partially update post fields
      operationId: patchPost
      tags: [Posts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                caption: { type: string }
                location: { type: string }
                tags: { type: array, items: { type: string } }
      responses:
        '200':
          description: Post patched
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete a post
      operationId: deletePost
      tags: [Posts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Post deleted successfully (No Content)
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/posts/{id}/comments:
    get:
      summary: Get comments for a post
      tags: [Comments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Comments list
        '404':
          $ref: '#/components/responses/NotFoundError'

    post:
      summary: Add comment to post
      tags: [Comments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateInput'
      responses:
        '201':
          description: Comment created
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/posts/{id}/likes:
    post:
      summary: Toggle like status on a post
      tags: [Posts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated like state
  /api/v1/services/soap-call:
    get:
      summary: Read WSDL and perform live SOAP call
      tags: [Services]
      parameters:
        - name: city
          in: query
          schema: { type: string, default: Boston }
      responses:
        '200':
          description: SOAP call execution result
  /api/v1/services/circuit-breaker:
    get:
      summary: Test Circuit Breaker with states (CLOSED, OPEN, HALF_OPEN)
      tags: [Services]
      parameters:
        - name: fail
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Circuit closed or healthy
        '503':
          description: Circuit open with fallback
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        name: { type: string }
        avatar: { type: string }
        bio: { type: string }
        location: { type: string }
        followersCount: { type: integer }
        followingCount: { type: integer }
        postsCount: { type: integer }
        verified: { type: boolean }
    PostCreateInput:
      type: object
      required: [caption, imageUrl]
      properties:
        caption: { type: string, minLength: 1 }
        imageUrl: { type: string }
        location: { type: string }
        eventTime: { type: string }
        capacity: { type: string }
        timeRemaining: { type: string }
        tags: { type: array, items: { type: string } }
    CommentCreateInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
    Post:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        author:
          $ref: '#/components/schemas/User'
        caption: { type: string }
        imageUrl: { type: string }
        location: { type: string }
        eventTime: { type: string }
        capacity: { type: string }
        timeRemaining: { type: string }
        likesCount: { type: integer }
        commentsCount: { type: integer }
        isLiked: { type: boolean }
        isSaved: { type: boolean }
        tags: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
    PostsFeedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        pagination:
          type: object
          properties:
            limit: { type: integer }
            cursor: { type: string, nullable: true }
            nextCursor: { type: string, nullable: true }
            hasMore: { type: boolean }
            count: { type: integer }
            total: { type: integer }
        links:
          type: object
          properties:
            self: { type: string }
            next: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, timestamp, correlationId]
          properties:
            code: { type: string, example: NOT_FOUND }
            message: { type: string, example: Resource was not found. }
            details: { nullable: true }
            timestamp: { type: string, format: date-time }
            correlationId: { type: string, example: 550e8400-e29b-41d4-a716-446655440000 }
  responses:
    NotModified:
      description: 304 Not Modified - Resource has not changed based on If-None-Match ETag
    BadRequestError:
      description: 400 Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundError:
      description: 404 Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: 422 Unprocessable Entity - JSON Schema validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitError:
      description: 429 Too Many Requests
      headers:
        X-RateLimit-Limit: { schema: { type: integer } }
        X-RateLimit-Remaining: { schema: { type: integer } }
        X-RateLimit-Reset: { schema: { type: integer } }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'